Fix previous hard state

+++ configobject/objecttypes/service/servicestate/servicestate.go
@@ -47,7 +47,7 @@ type ServiceState struct {
 	StateType					float32	`json:"state_type"`
 	SoftState					float32	`json:"state"`
 	HardState					float32	`json:"NONE"` //TODO (NoH): I'm not sure where to get this from
-	PreviousHardState 			uint8 	`json:"previous_hard_state"`
+	PreviousHardState 			float32	`json:"previous_hard_state"`
 	Attempt						float32	`json:"check_attempt"`
 	Severity					float32	`json:"severity"`
 	Output						string	`json:"output"`
